datatype changes

+++ Backend/order/src/main/java/com/urbanstyle/order/Contoller/VendorDashboardController.java
@@ -45,7 +45,7 @@ public class VendorDashboardController {
 	 * @return LAST 5 ORDERS
 	 */
 	@RequestMapping(value= {"/getLastOrdersForVendor"},method= {RequestMethod.POST,RequestMethod.GET})
-	public Map<String,Object> getLastOrdersForVendor(@RequestParam(value="offset")int offset,@RequestParam(value="vendorId")int vendorId,HttpServletRequest request,
+	public Map<String,Object> getLastOrdersForVendor(@RequestParam(value="offset")int offset,@RequestParam(value="vendorId")long vendorId,HttpServletRequest request,
 			@RequestParam(value="status")String status,HttpServletResponse response){
 		Map<String, Object> resultMap = new HashMap<String,Object>();
 		resultMap.put("lastOrders",orderService.getLastOrdersForVendor(offset,vendorId,status));
